🖌️ User settings polishing
